web app/requirements.txt
Flask==3.1.0
numpy==2.2.1
pandas==2.2.3
pytz==2024.2
Requests==2.32.3
xgboost==2.1.3
python-dotenv==0.21.0